+++ docs/changelog.md
@@ -4,6 +4,11 @@ All notable changes to Homeboy CLI are documented in this file.
 
 (This file is embedded into the CLI binary and is also viewable via `homeboy changelog`.)
 
+## Unreleased
+
+### Added
+- Add --project/-p flag to deploy command for explicit project specification
+
 ## [0.30.15] - 2026-01-20
 
 ### Added